feat(nuxt.config): sitemap configuration in nuxt.config
This commit changes the structure of the sitemap object in nuxt.config file to add error handling for failed API calls. Previously, a failed fetch to `${API_BASE_URL}/sitemap` would cause program to crash. Now, if the fetch fails, an empty array is returned to prevent any interruption. This adds an extra layer of reliability to sitemap generation process.