feat: enhance UX and functionality
In this commit, the page was updated to provide a more interactive and user-friendly experience to users.

There are changes in how the albums are fetched which provides server-side load distribution. Also, a lazy-loading mechanism is introduced for an optimized image and album load which improves page load speed.

The refactor also adds GSAP animations giving the website a cleaner look, and also enhances the gallery and film sections significantly with the addition of a "card" style for displaying content, which improves overall aesthetics.

Moreover, each film now has their separate card leading to their Vimeo links for a much better UX. Improvements in the dark-mode are also made including changes in the color scheme for better visibility and contrast.